+++ spec/functions.spec.js
@@ -1,5 +1,7 @@
 describe('basic functionality', function() {
 
+  //= functions/functions.CI.spec.js
+
   //= functions/functions.setlng.spec.js
 
   //= functions/functions.preload.spec.js