+++ test/test.js
@@ -753,6 +753,29 @@ describe('i18next', function() {
   });
   describe('basic functionality', function() {
   
+    describe('CI mode', function() {
+    
+      beforeEach(function(done) {
+        i18n.init(i18n.functions.extend(opts, {
+          resStore: {
+            'en-US': { translation: { 'simpleTest': 'ok_from_en-US' } },
+            'de-DE': { translation: { 'simpleTest': 'ok_from_de-DE' } }
+          }
+        }), function(t) { done(); } );
+      });
+    
+      it('it should provide resources for set language', function(done) {
+        expect(i18n.t('simpleTest')).to.be('ok_from_en-US');
+    
+        i18n.setLng('CIMode', function(t) {
+            expect(t('simpleTest')).to.be('simpleTest');
+            done();
+        });
+    
+      });
+    
+    });
+  
     describe('setting language', function() {
     
       beforeEach(function(done) {