+++ releasenotes.md
@@ -5,6 +5,7 @@
 - corrected latvian plurals based on [issue 231](https://github.com/jamuhl/i18next/issues/231)
 - allow array of fallback languages
 - allow int in values passed to shortcut sprintf
+- setLng to 'cimode' will trigger a CI mode returning 'key' instead of translation
 
 ### 1.7.2
 - introducing option fallbackOnEmpty -> empty string will fallback